chore/client: move Spinner for global component
